+++ cmp-common/src/commonMain/kotlin/ru/beryukhov/coffeegram/view/CalendarTable.kt
@@ -1,24 +1,30 @@
 package ru.beryukhov.coffeegram.view
 
+import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.lazy.grid.GridCells
+import androidx.compose.foundation.lazy.grid.LazyVerticalGrid
+import androidx.compose.material3.HorizontalDivider
+import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.platform.testTag
 import androidx.compose.ui.text.AnnotatedString
 import androidx.compose.ui.text.ParagraphStyle
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
-import com.slapps.cupertino.adaptive.AdaptiveHorizontalDivider
-import kotlinx.collections.immutable.PersistentList
 import kotlinx.collections.immutable.PersistentMap
-import kotlinx.collections.immutable.toPersistentList
 import kotlinx.collections.immutable.toPersistentMap
-import kotlinx.datetime.DayOfWeek
+import kotlinx.datetime.LocalDate
 import kotlinx.datetime.Month
 import org.jetbrains.compose.ui.tooling.preview.Preview
 import ru.beryukhov.coffeegram.app_ui.CoffeegramTheme
@@ -26,31 +32,29 @@ import ru.beryukhov.coffeegram.data.CoffeeTypes
 import ru.beryukhov.coffeegram.data.Picture
 import ru.beryukhov.date_time_utils.YearMonth
 import ru.beryukhov.date_time_utils.dateFormatSymbolsShortWeekdays
-import ru.beryukhov.date_time_utils.getShortDisplayName
 
 private data class DayItem(
     val day: String,
-    val coffeePicture: Picture,
+    val isToday: Boolean = false,
+    val coffeePicture: Picture = Picture.EMPTY,
     val dayOfMonth: Int? = null
-) {
-    companion object {
-        val EMPTY: DayItem = DayItem("", Picture.EMPTY)
-    }
-}
+)
 
 @Composable
 private fun DayCell(
     dayItem: DayItem,
-    onClick: () -> Unit,
     modifier: Modifier = Modifier,
-    ) {
+    onClick: (() -> Unit)? = null
+) {
     Column(
         horizontalAlignment = Alignment.CenterHorizontally,
-        modifier = if (dayItem.dayOfMonth == null) {
-            modifier
-        } else {
-            modifier.clickable(onClick = onClick)
-        }
+        modifier = modifier
+            .clickable(
+                enabled = onClick != null,
+                onClick = onClick ?: {}
+            )
+            .background(if (dayItem.isToday) MaterialTheme.colorScheme.primaryContainer else Color.Transparent)
+            .testTag("Day")
     ) {
         with(dayItem) {
             coffeePicture(
@@ -66,105 +70,56 @@ private fun DayCell(
                     paragraphStyle = ParagraphStyle(textAlign = TextAlign.Center)
                 )
             )
+            HorizontalDivider()
         }
     }
 }
 
-@Composable
-private fun WeekRow(
-    dayItems: PersistentList<DayItem?>,
-    onClick: (dayOfMonth: Int) -> Unit,
-    modifier: Modifier = Modifier
-) {
-    val weekDaysItems = dayItems.toMutableList()
-    weekDaysItems.addAll(listOf(DayItem.EMPTY) * (7 - weekDaysItems.size))
-    Column(horizontalAlignment = Alignment.CenterHorizontally) {
-        Row(modifier = Modifier.fillMaxWidth()) {
-            for (dayItem in weekDaysItems) {
-                DayCell(
-                    dayItem = dayItem
-                        ?: DayItem.EMPTY,
-                    onClick = { dayItem?.dayOfMonth?.let { onClick(it) } },
-                    modifier = Modifier.weight(1f)
-                )
-            }
-        }
-        AdaptiveHorizontalDivider()
-    }
-}
-
-@Composable
-private fun MonthTableAdjusted(
-    weekItems: PersistentList<PersistentList<DayItem?>>,
-    onClick: (dayOfMonth: Int) -> Unit,
-    modifier: Modifier = Modifier
-) {
-    Column(horizontalAlignment = Alignment.CenterHorizontally, modifier = modifier) {
-        weekItems.map { WeekRow(dayItems = it, onClick) }
-    }
-}
-
-private class WeekDayVectorPair(
-    val day: Int,
-    val weekDay: DayOfWeek,
-    var coffeePicture: Picture,
-) {
-    fun toDayItem(): DayItem =
-        DayItem("$day", coffeePicture, day)
-}
-
-// todo - replace by app's MonthTable with more features and better table implementation
 @Composable
 fun MonthTable(
     yearMonth: YearMonth,
+    today: LocalDate,
     filledDayItemsMap: PersistentMap<Int, Picture>,
     onClick: (dayOfMonth: Int) -> Unit,
     modifier: Modifier = Modifier
 ) {
-    val weekDays: PersistentList<DayItem> = getWeekDaysNames().map { DayItem(it, Picture.EMPTY) }
-        .toPersistentList()
-    val days1to31 = mutableListOf<Int>()
-    for (i in 1..31) {
-        days1to31.add(i)
-    }
-    val days = days1to31.filter { yearMonth.isValidDay(it) }
-        .associateBy(
-            { it },
-            {
-                WeekDayVectorPair(
-                    it,
-                    yearMonth.atDay(it).dayOfWeek,
-                    Picture.EMPTY
-                )
+    Column(modifier = modifier) {
+        LazyVerticalGrid(
+            columns = GridCells.Fixed(7),
+            contentPadding = PaddingValues(horizontal = 8.dp),
+            horizontalArrangement = Arrangement.spacedBy(4.dp),
+        ) {
+            items(7) { index ->
+                Column(horizontalAlignment = Alignment.CenterHorizontally, modifier = Modifier.padding(top = 8.dp)) {
+                    Text(
+                        text = getWeekDaysNames()[index],
+                        textAlign = TextAlign.Center,
+                    )
+                    HorizontalDivider()
+                }
             }
-        )
-        .toMutableMap()
-    filledDayItemsMap.forEach { days[it.key]?.coffeePicture = it.value }
-    val weekDaysStrings = getWeekDaysNames()
-    val numberOfFirstDay = weekDaysStrings.indexOf(
-        days[1]!!.weekDay.getShortDisplayName()
-    )
-    val daysList: List<WeekDayVectorPair> = days.toList().sortedBy { it.first }.map { it.second }
-    val firstWeek: PersistentList<DayItem> =
-        (listOf(DayItem.EMPTY) * numberOfFirstDay + daysList.take(7 - numberOfFirstDay)
-            .map(WeekDayVectorPair::toDayItem)).toPersistentList()
-
-    val secondToSixWeeks: List<PersistentList<DayItem>> = listOf(2, 3, 4, 5, 6).map {
-        daysList.drop(7 * (it - 1) - numberOfFirstDay).take(7)
-    }.filterNot { it.isEmpty() }
-        .map { it.map(WeekDayVectorPair::toDayItem).toPersistentList() }
-
-    val weekItems = mutableListOf(
-        weekDays,
-        firstWeek
-    )
-    weekItems.addAll(secondToSixWeeks)
-
-    return MonthTableAdjusted(
-        weekItems.toPersistentList(),
-        onClick,
-        modifier = modifier
-    )
+        }
+        val itemsOffset = yearMonth.atDay(1).dayOfWeek.ordinal + 1
+        val daysInMonth = yearMonth.lengthOfMonth()
+        LazyVerticalGrid(
+            columns = GridCells.Fixed(7),
+            contentPadding = PaddingValues(8.dp),
+            horizontalArrangement = Arrangement.spacedBy(4.dp),
+        ) {
+            items(itemsOffset + daysInMonth) { index ->
+                val day = index - itemsOffset
+                if (day >= 0) {
+                    val dayItem = DayItem(
+                        day = (day + 1).toString(),
+                        isToday = yearMonth.atDay(day + 1) == today,
+                        coffeePicture = filledDayItemsMap[day + 1] ?: Picture.EMPTY,
+                        dayOfMonth = day + 1
+                    )
+                    DayCell(dayItem = dayItem, onClick = { onClick(day + 1) })
+                }
+            }
+        }
+    }
 }
 
 @Preview
@@ -178,19 +133,12 @@ private fun TablePreview() {
 @Composable
 fun SampleTable(modifier: Modifier = Modifier) =
     MonthTable(
-        YearMonth(2020, Month.JULY),
-        mapOf(2 to CoffeeTypes.Cappuccino.icon).toPersistentMap(),
+        yearMonth = YearMonth(2020, Month.JULY),
+        today = LocalDate(2020, 7, 14),
+        filledDayItemsMap = mapOf(2 to CoffeeTypes.Cappuccino.icon).toPersistentMap(),
         modifier = modifier,
         onClick = {},
     )
 
 fun getWeekDaysNames(): List<String> =
     dateFormatSymbolsShortWeekdays()
-
-private operator fun <E> List<E>.times(i: Int): List<E> {
-    val result = mutableListOf<E>()
-    repeat(i) {
-        result.addAll(this)
-    }
-    return result
-}
